Separate the set of general projects from the Scala artifacts
This commit refactors the 2.11.x community file, by placing
into a common file all the projects that do not depend upon
the way in which the Scala artifacts and modules are generated.

By doing so, it becomes possible to test the same set of
projects using, for example, an optimized or non-optimized
compiler, or the GenBCode backend, or different Scala
options that should be used for all projects (to be specified
using using "vars.base").