Testing a combination of flags that should help with permgen errors
There must be a permgen leak in sbt that is somehow exercised by
dbuild; I could get permgen errors in sbt alone, although over a
much longer test.

While launching sbt, dbuild uses "-XX:+CMSClassUnloadingEnabled";
however, that option is not effective unless the additional
option "-XX:+UseConcMarkSweepGC" is also used. This commit adds
that option manually, as a test: in theory, the combination of
the two should have a beneficial effect on PermGen troubles.

If that works, I will add it as a default option in dbuild.