Upgrade to Scalaz 7.1.x series.
This is the latest stable release of Scalaz. We upgrade it only in the
default space. There's another scalaz defined as "scalaz-210" and is used
for building sbt plugins. We keep it at 7.0.x to avoid problems with
sbteclipse. Once sbteclipse upgrades to latest scalaz, we should move
that version in our config too.