Twirl has been moved to playframework
Latest version of Play depends on a new twirl that has been moved to
Play organization on GitHub. That version doesn't build in dbuild due to
pattern matching on scalaBinaryVersion, see playframework/twirl#45 and
typesafehub/dbuild#145 for details. For that reason, we are building Play's
Twirl from a head of PR which includes workaround for scalaBinaryVersion
problem.

However, we can't switch over to Play's Twirl altogether just yet. The
reason is that spray fork depends on old twirl.

The good news is that Spray's and Play's Twirl do not share group ids so
we can just build both of them and they'll be picked up by projects that
need them. That's what we are doing here.