Remove references to sbt 0.12.x.
They are not needed because sbt 0.13.x is backwards compatible.

Switch to custom fork of sbt-idea plugin. That version of sbt-idea
depends on newer version of sbt-pgp plugin. Newer version has correct
dependencies and fixes the following problem:

[sbt-idea] [error] Modules were resolved with conflicting cross-version suffixes in {file:/localhome/jenkinsdbuild/workspace/Community-2.11.x-Java8/target-0.9.1/extraction/082c36b94bc0dd64110a403174908c48bd1a3558/projects/1beff2840ad3d104b774ff8f785a34e2ccfeccfb/project/}root-1beff2840ad3d104b774ff8f785a34e2ccfeccfb-build:
[sbt-idea] [error]    net.databinder:dispatch-core _2.10, _2.9.1
[sbt-idea] [error]    net.databinder:dispatch-http _2.10, _2.9.1
[sbt-idea] [error]    net.databinder:dispatch-futures _2.10, _2.9.1
[sbt-idea] java.lang.RuntimeException: Conflicting cross-version suffixes in: net.databinder:dispatch-core, net.databinder:dispatch-http, net.databinder:dispatch-futures

We also upgrade to latest release of android-sdk-plugin. This is required as part
of the update of sbt-idea plugin.

Those changes pave the way for resolving #83.