Various fixes to restore the community build
Upgrade sbt to 0.13.8. This fixes a recent update to the npm build
definition:
https://github.com/typesafehub/npm/commit/16c2a559b1d003efc298356941915b19b2ff1d38#commitcomment-10754486

For some projects we have to use 0.13.7 for now, comments in the file.

Adds a temporary repository to the resolvers to enable the fix for
https://github.com/sbt/sbt-web/pull/107/files

Use a fixed reference for sbt-js-engine to avoid issues with the
bintray publishing sbt plugin.
https://github.com/sbt/sbt-js-engine/commit/f28889c9c2a6581d1d7a1e530cb5cd71cb0e7d9c#commitcomment-10833306