fix crosstalk between jdk6 and jdk8 builds
fixes #173

I had previously committed this as 77f26d4, then reverted it
later because I wasn't sure it had any effect, and because I wasn't
sure the problem I was trying to fix even existed.  as documented
in #173, the problem definitely exists, so let's try again and see
if this fixes it.