keep builds for a few days longer
I believe, based on my monitoring of the disk space situation on the
behemoths over the past few months, that this will be fine, and that
it will save us time during maintenance since we won't have to wait as
often for discarded builds to be redone.

I will continue monitoring the disk space to be sure.