move specs2 to 3.8 (still forked though) and remove scalaz-stream
we have fs2 instead, and specs2 no longer depends on scalaz-stream

specs2 fork is to fix a `packagedArtifacts` setting

having dropped specs2 2.x we must fork scalamock

also remove spray-twirl (which wanted specs2 2.x). no commits since
2014. the action is now in play-twirl instead. currently nobody
depends on this, but that could be because spray itself is currently
commented out? let's cross that bridge when we come to it.
https://github.com/spray/spray/issues/887 seems to indicate it
won't be a problem