freeze all SHAs temporarily
also removed ScalaMeter because it is currently known to hang;
that's https://github.com/scala/scala-dev/issues/280
and a fix is in the works

until 2.12.1 is out the door, we don't want new errors constantly
popping up at us, especially now that the community build is part
of the release-main flow in Jenkins.  so freeze all SHAs for now.
to be unfrozen after the release is out the door.

doing this on 2.12.x, not by making a 2.12.1 branch, since the release
flow is configured to use the same branch in all involved repos.

note that I have *not* yet taken steps to make sure these SHAs won't
eventually disappear from GitHub, as detailed at
https://github.com/scala/community-builds/issues/338 . I should
do that at the time we unfreeze after 2.12.1.