remove scalafmt workarounds
latest scalafmt versions are more dbuild-friendly.

scala-json-ast already upgraded scalafmt themselves.

akka-sse was still on an old version.  we are already forking akka-sse,
so I refreshed the fork and added a new commit with the scalafmt upgrade.
I also separately submitted that commit as a PR.  (even if it's accepted,
we'll still need to stay forked, because sbt-bintray.)