add coursier, pureconfig, circe-config (and scalacheck-shapeless)
the parts of this are not really related to each other

this also contains a not-quite-finished attempt to add more ammonite
here, with the final crucial bit still commented put.  (upickle is
reinstated as part of the prep for that.)  its only relation to the
rest of this is that coursier is a dependency of ammonite's, so
having coursier in made the attempt possible.