shuffle project order, improving usability/maintainability
in general, we would like to order the projects so that every project
always comes after all its dependencies.  this makes it easy to save
time during dependency extraction in test runs by chopping out
everything below the project you care about.

this commit doesn't achieve that for the whole config file,
but it does fix a number of such ordering problems.  in
general, we are trying to list keystone projects earlier,
gravy later.