Pin some projects to sbt 1.3.13
Most of the projects in the community build pass with the latest sbt
1.4.x code. There are a few exceptions and this commit pins those
projects to use 1.3.13. For all droste, lagom, and playframework, there
are outstanding PRs to make the builds 1.4.x compatible. Enumeratum
doesn't work because scala 2.12.11 broke binary compatibility with a
plugin it uses (I have a PR for the bin compat but it's more involved to
fix the plugin to make it 2.12.x compatible for all x). Zinc doesn't
work because the pinned version uses a version of contraband that isn't
2.12.12 compatible and it can't easily be upgraded because of weirdness
with droste and the protobuf plugin.